+++ app/Views/Members/Member-Account-Sidebar.php
@@ -16,6 +16,7 @@ use Libs\Language;
     </div>
     <ul class='list-group list-group-flush'>
       <li class='list-group-item'><a href='<?=DIR?>Edit-Profile' rel='nofollow'><?=Language::show('mem_act_edit_profile', 'Members'); ?></a></li>
+      <li class='list-group-item'><a href='<?=DIR?>Edit-Profile-Images' rel='nofollow'><?=Language::show('mem_act_edit_profile_images', 'Members'); ?></a></li>
       <li class='list-group-item'><a href='<?=DIR?>Change-Email' rel='nofollow'><?=Language::show('mem_act_change_email', 'Members'); ?></a></li>
       <li class='list-group-item'><a href='<?=DIR?>Change-Password' rel='nofollow'><?=Language::show('mem_act_change_pass', 'Members'); ?></a></li>
       <li class='list-group-item'><a href='<?=DIR?>Privacy-Settings' rel='nofollow'><?=Language::show('mem_act_privacy_settings', 'Members'); ?></a></li>